:truck: api: move with-required-action-session serializer from lib to src
