+++ api/src/certification/session-management/infrastructure/serializers/with-required-action-session-serializer.js
